[CAZ-2533] Show choice to pay via Direct Debit if user set one up (#253)
