+++ .drone.yml
@@ -139,8 +139,7 @@ steps:
     pull: never
     commands:
       - bundle audit check --update
-      # to exclude id from audit(e.g. 961) just add `--exclude 961`
-      - yarn run improved-yarn-audit --exclude '1488, 1179' --retry-on-network-failure
+      - yarn run improved-yarn-audit --retry-on-network-failure
       - brakeman
     volumes:
       - name: docker_sock
@@ -361,8 +360,7 @@ steps:
     pull: never
     commands:
       - bundle audit check --update
-      # to exclude id from audit(e.g. 961) just add `--exclude 961`
-      - yarn run improved-yarn-audit --exclude '1488, 1179' --retry-on-network-failure
+      - yarn run improved-yarn-audit --retry-on-network-failure
       - brakeman
     volumes:
       - name: docker_sock
@@ -608,8 +606,7 @@ steps:
     pull: never
     commands:
       - bundle audit check --update
-      # to exclude id from audit(e.g. 961) just add `--exclude 961`
-      - yarn run improved-yarn-audit --exclude '1488, 1179' --retry-on-network-failure
+      - yarn run improved-yarn-audit --retry-on-network-failure
       - brakeman
     volumes:
       - name: docker_sock
@@ -861,8 +858,7 @@ steps:
     pull: never
     commands:
       - bundle audit check --update
-      # to exclude id from audit(e.g. 961) just add `--exclude 961`
-      - yarn run improved-yarn-audit --exclude '1488, 1179' --retry-on-network-failure
+      - yarn run improved-yarn-audit --retry-on-network-failure
       - brakeman
     volumes:
       - name: docker_sock
