+++ Gemfile
@@ -24,7 +24,7 @@ gem 'turbolinks'
 gem 'webpacker'
 
 group :development, :test do
-  gem 'byebug', platforms: %i[mri mingw x64_mingw]
+  gem 'byebug'
   gem 'dotenv-rails'
   gem 'haml-rails'
   gem 'rspec-rails'
