+++ features/debits.feature
@@ -20,23 +20,6 @@ Feature: Debits
       And I press 'Confirm payment' button
     Then I should see success message
 
-  Scenario: Making a Direct Debit payment with no active mandate
-    When I have only inactive mandates for selected CAZ
-      And I visit the make payment page
-      And I press the Continue
-    Then I select 'Birmingham'
-      And I press the Continue
-    Then I click Next 7 days tab
-      And I select any date for vrn on the payment matrix
-      And I press the Continue
-    Then I want to confirm my payment
-      And I press the Continue
-    Then I select 'Direct Debit'
-      And I press the Continue
-    Then I should see 'First create a Direct Debit mandate'
-      And I press the Continue
-    Then I should see 'Your Direct Debits'
-
   Scenario: Visiting the manage Direct Debit page with no mandates
     When I have no mandates
       And I visit the manage Direct Debit page
