+++ package.json
@@ -6,10 +6,10 @@
     "@rails/actioncable": "^6.0.2",
     "@rails/activestorage": "^6.0.2",
     "@rails/ujs": "^6.0.2",
-    "@rails/webpacker": "^5.0.1",
+    "@rails/webpacker": "^5.1.1",
     "core-js": "^3.6.4",
     "govuk-frontend": "^3.6.0",
-    "improved-yarn-audit": "^1.2.0",
+    "improved-yarn-audit": "^2.0.0",
     "rails-erb-loader": "^5.5.2",
     "regenerator-runtime": "^0.13.5"
   },
